requirements.txt
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
exceptiongroup==1.2.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
kaitaistruct==0.10
outcome==1.3.0.post0
pyasn1==0.6.0
pycparser==2.22
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
requests==2.31.0
selenium==4.19.0
selenium-wire==5.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
zstandard==0.22.0
python-dotenv~=1.1.0
boto3